+++ app/src/main/java/it/niedermann/nextcloud/deck/persistence/sync/adapters/db/dao/widgets/filter/FilterWidgetDao.java
@@ -12,5 +12,5 @@ public interface FilterWidgetDao extends GenericDao<FilterWidget> {
     void delete (Long filterWidgetId);
 
     @Query("SELECT * FROM FilterWidget where id = :filterWidgetId")
-    FilterWidget getFilterWidgetByIdDirectly(Long filterWidgetId);
+    FilterWidget getFilterWidgetByIdDirectly(Integer filterWidgetId);
 }
