+++ app/src/main/java/it/niedermann/nextcloud/deck/persistence/sync/adapters/db/dao/widgets/filter/FilterWidgetSortDao.java
@@ -11,8 +11,8 @@ import it.niedermann.nextcloud.deck.persistence.sync.adapters.db.dao.GenericDao;
 @Dao
 public interface FilterWidgetSortDao extends GenericDao<FilterWidgetSort> {
     @Query("DELETE FROM FilterWidgetSort WHERE filterWidgetId = :filterWidgetId")
-    void deleteByFilterWidgetId (Long filterWidgetId);
+    void deleteByFilterWidgetId (Integer filterWidgetId);
 
     @Query("select * FROM FilterWidgetSort WHERE filterWidgetId = :filterWidgetId order by ruleOrder asc")
-    List<FilterWidgetSort> getFilterWidgetSortByFilterWidgetIdDirectly(Long filterWidgetId);
+    List<FilterWidgetSort> getFilterWidgetSortByFilterWidgetIdDirectly(Integer filterWidgetId);
 }
